refactor: Rename files to follow consistent [entity].[method].ts pattern
