+++ src/api.ts
@@ -2,8 +2,8 @@ import express from 'express'
 import pino from 'pino-http'
 import swaggerJsdoc from 'swagger-jsdoc'
 import { apiReference } from '@scalar/express-api-reference'
-import readCompanies from './routes/readCompanies'
-import updateCompanies from './routes/updateCompanies'
+import readCompanies from './routes/companies.read'
+import updateCompanies from './routes/companies.update'
 import { errorHandler } from './routes/middlewares'
 import { swaggerOptions } from './swagger'
